[feat] 네이밍 변경 DemoSessionDelegate => StarlinkDelegate
