+++ Cllog/Projects/AppModule/Starlink/Sources/Configure/Configure.swift
@@ -17,7 +17,7 @@ public protocol StarlinkConfiguration: Sendable {
 
 extension StarlinkConfiguration {
     public var queue: OperationQueue? { nil }
-    public var delegate: URLSessionDelegate? { DemoSessionDelegate() }
+    public var delegate: URLSessionDelegate? { StarlinkDelegate() }
 }
 
 public extension Starlink {
@@ -32,19 +32,5 @@ public extension Starlink {
             configuration.timeoutIntervalForResource = 60
             return configuration
         }
-    }
-}
-
-private final class DemoSessionDelegate: NSObject, URLSessionDelegate, URLSessionDataDelegate {
-    func urlSession(_ session: URLSession, dataTask: URLSessionDataTask, didReceive data: Data) {
-        NetworkLogger.shared.logDataTask(dataTask, didReceive: data)
-    }
-    
-    func urlSession(_ session: URLSession, task: URLSessionTask, didFinishCollecting metrics: URLSessionTaskMetrics) {
-        NetworkLogger.shared.logTask(task, didFinishCollecting: .init(metrics: metrics))
-    }
-    
-    func urlSession(_ session: URLSession, task: URLSessionTask, didCompleteWithError error: (any Error)?) {
-        NetworkLogger.shared.logTask(task, didCompleteWithError: error)
     }
 }
